requirements.txt
customtkinter
tk
pyinstaller
dolphin-memory-engine-mpn
json5
requests
pywin32
Pillow
flask
websocket-client